moves modal folder
